airflow_pipeline/requirements.txt
python-dotenv
psycopg2-binary
requests
beautifulsoup4
chardet
boto3
pymilvus==2.5.0
openai
unidecode
langchain-openai
langchain-community
openai
python-docx
mammoth
openpyxl
pymupdf